스케줄링 1초 단위로 맞춤

+++ src/pyodide-simulator/models.py
@@ -19,7 +19,7 @@ class ExecutionBlock:
 @dataclass
 class ProcessMetric:
     pid: str
-    bt: float
+    bt: int
     at: int
     tt: float
     wt: float
